requirements_dashboard.txt
# Dashboard-specific requirements for LMArena Bridge
# Install with: pip install -r requirements_dashboard.txt

# Core dependencies (if not already installed from main requirements.txt)
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.2

# Additional dependencies for dashboard
requests>=2.31.0
